- Improve DBI compliance (#1, #5, #10, #15, #17, #40).

+++ src/MariaRow.h
@@ -33,8 +33,8 @@ public:
   SEXP valueString(int j);
   SEXP valueRaw(int j);
   double valueDateTime(int j);
-  int valueDate(int j);
-  int valueTime(int j);
+  double valueDate(int j);
+  double valueTime(int j);
 
   void setListValue(SEXP x, int i, int j);
